Components: Extract BlockToolbar as a reusable component
