+++ editor/components/block-mover/index.js
@@ -15,9 +15,9 @@ import { getBlockType } from '@wordpress/blocks';
  * Internal dependencies
  */
 import './style.scss';
-import { isFirstBlock, isLastBlock, getBlockIndex, getBlock } from '../selectors';
 import { getBlockMoverLabel } from './mover-label';
-import { selectBlock } from '../actions';
+import { isFirstBlock, isLastBlock, getBlockIndex, getBlock } from '../../selectors';
+import { selectBlock } from '../../actions';
 
 export function BlockMover( { onMoveUp, onMoveDown, isFirst, isLast, uids, blockType, firstIndex } ) {
 	// We emulate a disabled state because forcefully applying the `disabled`
