+++ editor/components/block-mover/style.scss
@@ -1,19 +1,3 @@
-.editor-block-mover {
-	position: absolute;
-	top: 2px;
-	left: 4px;
-	height: $text-editor-font-size * 4; // same height as an empty paragraph
-	padding: 4px 14px 6px 0; // handles hover area
-	z-index: z-index( '.editor-block-mover' );
-
-	// Mobile, to be revisited
-	display: none;
-
-	@include break-small {
-		display: block;
-	}
-}
-
 .editor-block-mover__control {
 	display: block;
 	padding: 0;
