+++ editor/components/block-mover/test/mover-label.js
