+++ editor/components/block-settings-menu/block-delete-button.js
@@ -13,7 +13,7 @@ import { IconButton } from '@wordpress/components';
 /**
  * Internal dependencies
  */
-import { removeBlocks } from '../actions';
+import { removeBlocks } from '../../actions';
 
 export function BlockDeleteButton( { onDelete, onClick = noop, small = false } ) {
 	const label = __( 'Delete' );
