+++ editor/components/block-settings-menu/unknown-converter.js
@@ -13,8 +13,8 @@ import { getUnknownTypeHandlerName, rawHandler, serialize } from '@wordpress/blo
 /**
  * Internal dependencies
  */
-import { getBlock } from '../selectors';
-import { replaceBlocks } from '../actions';
+import { getBlock } from '../../selectors';
+import { replaceBlocks } from '../../actions';
 
 export function UnknownConverter( { block, convertToBlocks, small } ) {
 	if ( ! block || getUnknownTypeHandlerName() !== block.name ) {
