+++ editor/components/block-switcher/style.scss
