+++ editor/components/block-toolbar/style.scss
@@ -37,8 +37,6 @@
 
 	.editor-block-mover {
 		display: inline-flex;
-		height: auto;
-		position: initial;
 		padding: 0 5px;
 		align-items: stretch;
 	}
