+++ editor/modes/visual-editor/block-contextual-toolbar.js
@@ -12,7 +12,7 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import NavigableToolbar from '../../navigable-toolbar';
-import BlockToolbar from '../../block-toolbar';
+import { BlockToolbar } from '../../components';
 import { isFeatureActive } from '../../selectors';
 
 function BlockContextualToolbar( { hasFixedToolbar } ) {
